requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.10.0
streamlit>=1.25.0
polygon-api-client>=1.0.0
psycopg2-binary>=2.9.0
